Fix inconsistent DLL Linkage on MSVC compilers
A circular dependency between DREAM3DLib and OrientationLib existed with the StatsGen, Texture and OrientationOps headers. These
files have been moved to the OrientationLib library. All unnecessary #includes have been removed from headers in the context
of OrientationOps.

Removing the include of Hexagonal.cpp file from the RenumberFeaturesBinnedOrientations.cpp file. This is VERY bad to do and should
NEVER be done.

Updates #263

+++ Source/Plugins/OrientationAnalysis/OrientationAnalysisFilters/WritePoleFigure.cpp
@@ -345,7 +345,7 @@ void WritePoleFigure::execute()
   {
     // First find out how many voxels we are going to have. This is probably faster to loop twice than to
     // keep allocating memory everytime we find one.
-    for(int64_t i = 0; i < numPoints; ++i)
+    for(size_t i = 0; i < numPoints; ++i)
     {
       if (m_CellPhases[i] == phase)
       {
